refactor: Make it easier for future multiversioning
